+++ src/components/app/pageVoterGuide/constants.ts
@@ -18,6 +18,8 @@ type VoterGuideStep = {
   campaignName: UserActionCampaignName
 }
 
+export const ANALYTICS_NAME_USER_ACTION_FORM_GET_INFORMED = 'User Action Form Get Informed'
+
 export const VOTER_GUIDE_STEPS: VoterGuideStep[] = [
   {
     title: 'Get Informed',
