+++ src/components/app/pageVoterGuide/formConfig.ts
@@ -3,7 +3,7 @@ import { object, type z } from 'zod'
 import { GetUserFullProfileInfoResponse } from '@/app/api/identified-user/full-profile-info/route'
 import { zodGooglePlacesAutocompletePrediction } from '@/validation/fields/zodGooglePlacesAutocompletePrediction'
 
-export const FORM_NAME = 'Voter Guide - Address'
+export const FORM_NAME = 'Voter Guide - Get Informed'
 
 export const voterGuideFormValidationSchema = object({
   address: zodGooglePlacesAutocompletePrediction,
@@ -12,12 +12,12 @@ export const voterGuideFormValidationSchema = object({
 export type VoterGuideFormValues = z.infer<typeof voterGuideFormValidationSchema>
 
 export function getDefaultValues({ user }: { user?: GetUserFullProfileInfoResponse['user'] }) {
-  return {
-    address: user?.address
-      ? {
+  return user?.address
+    ? {
+        address: {
           description: user?.address.formattedDescription,
           place_id: user?.address.googlePlaceId,
-        }
-      : undefined,
-  }
+        },
+      }
+    : undefined
 }
