+++ src/components/app/pageVoterGuide/getInformedForm.tsx
@@ -1,0 +1,224 @@
+import {
+  CreateActionViewKeyRacesInput,
+  actionCreateUserActionViewKeyRaces,
+} from '@/actions/actionCreateUserActionViewKeyRaces'
+import { ANALYTICS_NAME_USER_ACTION_FORM_GET_INFORMED } from '@/components/app/pageVoterGuide/constants'
+import {
+  VoterGuideFormValues,
+  voterGuideFormValidationSchema,
+} from '@/components/app/pageVoterGuide/formConfig'
+import { KeyRacesSkeleton, KeyRacesList } from '@/components/app/pageVoterGuide/keyRacesList'
+import { UserActionFormLayout } from '@/components/app/userActionFormCommon'
+import { useRacesByAddress } from '@/components/app/userActionFormVoterAttestation/useRacesByAddress'
+import { dialogContentPaddingStyles } from '@/components/ui/dialog/styles'
+import { ErrorMessage } from '@/components/ui/errorMessage'
+import { FormField, FormItem, FormControl, Form } from '@/components/ui/form'
+import { GooglePlacesSelect } from '@/components/ui/googlePlacesSelect'
+import { useApiResponseForUserPerformedUserActionTypes } from '@/hooks/useApiResponseForUserPerformedUserActionTypes'
+import { useIsMobile } from '@/hooks/useIsMobile'
+import { convertAddressToAnalyticsProperties } from '@/utils/shared/sharedAnalytics'
+import { cn } from '@/utils/web/cn'
+import { trackFormSubmissionSyncErrors, triggerServerActionForForm } from '@/utils/web/formUtils'
+import { convertGooglePlaceAutoPredictionToAddressSchema } from '@/utils/web/googlePlaceUtils'
+import { identifyUserOnClient } from '@/utils/web/identifyUser'
+import {
+  catchUnexpectedServerErrorAndTriggerToast,
+  toastGenericError,
+} from '@/utils/web/toastUtils'
+import { zodResolver } from '@hookform/resolvers/zod'
+import { UserActionType } from '@prisma/client'
+import { ScrollArea } from '@radix-ui/react-scroll-area'
+import { useRouter } from 'next/navigation'
+import { useRef, useEffect } from 'react'
+import { useForm, useWatch } from 'react-hook-form'
+import * as Sentry from '@sentry/nextjs'
+import { Button } from '@/components/ui/button'
+import { useGoogleMapsScript } from '@/hooks/useGoogleMapsScript'
+
+interface GetInformedFormProps {
+  onFinish?: () => void
+  onUserActionCreated?: () => void
+  initialValues?: VoterGuideFormValues
+}
+
+export const GetInformedForm = (props: GetInformedFormProps) => {
+  const { onFinish, initialValues, onUserActionCreated } = props
+
+  const router = useRouter()
+  const scriptStatus = useGoogleMapsScript()
+
+  const { mutate } = useApiResponseForUserPerformedUserActionTypes()
+
+  const form = useForm<VoterGuideFormValues>({
+    defaultValues: initialValues,
+    resolver: zodResolver(voterGuideFormValidationSchema),
+  })
+
+  const isMobile = useIsMobile({ defaultState: true })
+  const inputRef = useRef<HTMLInputElement | null>(null)
+  const error = form.formState.errors?.address
+  const errorRef = useRef(error)
+
+  useEffect(() => {
+    if (!isMobile && !errorRef.current && !initialValues?.address) {
+      inputRef.current?.click()
+    }
+  }, [form, isMobile])
+
+  const address = useWatch({
+    control: form.control,
+    name: 'address',
+  })
+
+  const racesByAddressRequest = useRacesByAddress(address?.description, {
+    onError: e => {
+      if (e instanceof Error || e?.message) {
+        form.setError('address', {
+          message: e.message,
+        })
+      }
+    },
+    onSuccess: () => {
+      form.clearErrors('address')
+    },
+  })
+
+  const createViewKeyRacesAction = async () => {
+    const formValues = form.getValues()
+    if (!formValues.address) return
+
+    const addressSchema = await convertGooglePlaceAutoPredictionToAddressSchema(
+      formValues.address,
+    ).catch(e => {
+      Sentry.captureException(e)
+      catchUnexpectedServerErrorAndTriggerToast(e)
+      return null
+    })
+    if (!addressSchema) {
+      form.setError('address', {
+        message: 'Invalid address',
+      })
+      return
+    }
+    const payload: CreateActionViewKeyRacesInput = {
+      address: addressSchema,
+      usCongressionalDistrict: addressSchema?.usCongressionalDistrict,
+      usaState: addressSchema?.administrativeAreaLevel1,
+      shouldBypassAuth: true,
+    }
+    const result = await triggerServerActionForForm(
+      {
+        form,
+        formName: ANALYTICS_NAME_USER_ACTION_FORM_GET_INFORMED,
+        analyticsProps: {
+          ...(addressSchema ? convertAddressToAnalyticsProperties(addressSchema) : {}),
+          'User Action Type': UserActionType.VIEW_KEY_RACES,
+        },
+        payload,
+        onError: (_, e) => {
+          form.setError('address', {
+            message: e.message,
+          })
+          toastGenericError()
+        },
+      },
+      input =>
+        actionCreateUserActionViewKeyRaces(input).then(actionResult => {
+          if (actionResult && 'user' in actionResult && actionResult.user) {
+            identifyUserOnClient(actionResult.user)
+          }
+          return actionResult
+        }),
+    )
+    if (result.status === 'success') {
+      router.refresh()
+      void mutate()
+      onUserActionCreated?.()
+    }
+  }
+
+  useEffect(() => {
+    if (address?.description && scriptStatus === 'ready') {
+      void createViewKeyRacesAction()
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [address, scriptStatus])
+
+  return (
+    <UserActionFormLayout>
+      <div className="flex h-full flex-1 flex-col">
+        <ScrollArea className="min-h-[70vh] overflow-auto md:max-h-[70vh]">
+          <div className={cn(dialogContentPaddingStyles)}>
+            <div className="space-y-6">
+              <p className="text-center text-xl font-semibold">Get informed</p>
+              <p className="text-center text-fontcolor-muted">
+                See who's on your ballot and pledge to vote in this year's election.
+              </p>
+
+              <div className="mx-auto w-full max-w-lg">
+                <Form {...form}>
+                  <form
+                    id="view-key-races-form"
+                    onSubmit={form.handleSubmit(async () => {
+                      onFinish?.()
+                    }, trackFormSubmissionSyncErrors(ANALYTICS_NAME_USER_ACTION_FORM_GET_INFORMED))}
+                  >
+                    <FormField
+                      control={form.control}
+                      name="address"
+                      render={({ field }) => (
+                        <FormItem>
+                          <FormControl aria-invalid={!!error}>
+                            <GooglePlacesSelect
+                              {...field}
+                              className="h-14 rounded-full bg-secondary"
+                              onChange={newAddress => {
+                                form.clearErrors('address')
+                                field.onChange(newAddress)
+                              }}
+                              placeholder="Your full address"
+                              ref={inputRef}
+                              value={field.value}
+                            />
+                          </FormControl>
+                          {!!error && <ErrorMessage>{error?.message}</ErrorMessage>}
+                        </FormItem>
+                      )}
+                    />
+                  </form>
+                </Form>
+              </div>
+            </div>
+          </div>
+
+          <div className="space-y-6 md:space-y-10">
+            {racesByAddressRequest.isLoading ? (
+              <KeyRacesSkeleton />
+            ) : (
+              <KeyRacesList races={racesByAddressRequest?.data} />
+            )}
+          </div>
+        </ScrollArea>
+
+        <div
+          className="z-10 mt-auto flex flex-col items-center justify-center border border-t p-6 sm:flex-row md:px-12"
+          style={{ boxShadow: 'rgba(0, 0, 0, 0.2) 0px 1px 6px 0px' }}
+        >
+          <Button
+            className="ml-auto min-w-[130px]"
+            disabled={
+              form.formState.isSubmitting ||
+              racesByAddressRequest.isLoading ||
+              !form.formState.isValid
+            }
+            form="view-key-races-form"
+            size="lg"
+            type="submit"
+          >
+            I pledge to vote
+          </Button>
+        </div>
+      </div>
+    </UserActionFormLayout>
+  )
+}
